10-elastic-phoenix/requirements.txt
openai
python-dotenv
arize-phoenix
arize-phoenix-client
arize-phoenix-evals
elastic-opentelemetry
opentelemetry-instrumentation-httpx
openinference-instrumentation
openinference-instrumentation-openai
